+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.0 - 2020-10-18
+
+### Added
+- `isVariadic` method for variables.
+- `setIsVariadic` method for variables.
+- Implemented variadic support for variables in methods.
+
 ## 1.1.0 - 2020-10-17
 
 ### Added
@@ -44,5 +51,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The initial implementation of the package.
 
 # Versions
-- [1.1.0 > Unreleased](https://github.com/grizz-it/ast/compare/1.1.0...HEAD)
+- [1.2.0 > Unreleased](https://github.com/grizz-it/ast/compare/1.2.0...HEAD)
+- [1.1.0 > 1.2.0](https://github.com/grizz-it/ast/compare/1.1.0...1.2.0)
 - [1.0.0 > 2.0.0](https://github.com/grizz-it/ast/compare/1.0.0...1.1.0)